+++ core/src/jsMain/kotlin/nl/adaptivity/xmlutil/dom/Attr.kt
@@ -31,7 +31,7 @@ public actual external interface Attr : Node {
 
 public actual inline fun Attr.getNamespaceURI(): String? = namespaceURI
 public actual inline fun Attr.getPrefix(): String? = prefix
-public actual inline fun Attr.getLocalName(): String = localName
+public actual inline fun Attr.getLocalName(): String? = localName
 public actual inline fun Attr.getName(): String = name
 public actual inline fun Attr.getValue(): String = value
 public actual inline fun Attr.setValue(value: String) {