+++ core/src/nativeMain/kotlin/nl/adaptivity/xmlutil/dom/Attr.kt
@@ -32,7 +32,7 @@ public actual interface Attr : Node {
 
 public actual inline fun Attr.getNamespaceURI(): String? = namespaceURI
 public actual inline fun Attr.getPrefix(): String? = prefix
-public actual inline fun Attr.getLocalName(): String = localName
+public actual inline fun Attr.getLocalName(): String? = localName
 public actual inline fun Attr.getName(): String = name
 public actual inline fun Attr.getValue(): String = value
 public actual inline fun Attr.setValue(value: String) { this.value = value }